Support config fields with hyphens or camelCase
By default the example uses the same field name in the config file as the flag name. So a flag with a hyphenn is expected to a use a hyphen in the config file as well. For example:

--log-level=5 would be
```yaml
log-level: 5
```

I've added a behavior switch so that you can use camelCase in the config file instead, so that:

--log-level=5 would be
```yaml
logLevel: 5
```

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>